# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Food Delivery service API
  description: |-
    Toptal test project Food Delivery service API docs.
  version: '1.0'
servers:
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Forbidden:
      description: 'Forbidden'
    NotFound:
      description: 'Not Found'
paths:
...
  /block/{uuid}:
    post:
      tags: [Authentication]
      summary: 'Block a user'
      description: Blocks a user by UUID. For Only Moderators.
      parameters:
        - name: uuid
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User successfully blocked.
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
...